+++ rv/modules/gpio.py
@@ -11,9 +11,9 @@ class Gpio(Module):
     behaviors = {B.receives_audio, B.sends_audio}
 
     out = Controller(bool, False)
-    out_pin = Controller((0, 64), 0)
+    out_pin = Controller((0, 256), 0)
     out_threshold = Controller((0, 100), 50)
     in_ = Controller(bool, False)
-    in_pin = Controller((0, 64), 0)
+    in_pin = Controller((0, 256), 0)
     in_note = Controller((0, 128), 0)
     in_amplitude = Controller((0, 100), 100)