+++ rv/modules/kicker.py
@@ -22,7 +22,7 @@ class Kicker(Module):
     panning = Controller((-128, 128), 0)
     attack = Controller((0, 512), 0)
     release = Controller((0, 512), 32)
-    vol_addition = Controller((0, 1024), 0)
+    boost = Controller((0, 1024), 0)
     env_acceleration = Controller((0, 1024), 256)
     polyphony_ch = Controller((1, 4), 1)
     anticlick = Controller(bool, False)